Refactor dossier attachment components
